Added Robolectric tests along with basic ui + mocked data store

+++ app/src/main/java/me/syncify/unittestandroid/keynotes/KeyNote.java
@@ -4,8 +4,9 @@ package me.syncify.unittestandroid.keynotes;
  * Created by adarshpandey on 9/12/17.
  */
 
-public class KeyNote {
-    public int id;
+public class KeyNote
+{
+    public long id;
     public String title;
     public String desc;
